+++ CHANGELOG.md
@@ -17,6 +17,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Security
 
 
+## v4.0.3- 2017-11-21
+### Changed
+- 上映イベントの空席状況と、上映イベント自体のインポート期間を合わせるように調整。
+- MovieTheaterインポート時にidentifierフィールドがセットされるように修正。
+
+### Fixed
+- COAの認証エラーが頻出するバグ対応として[sskts-domain](https://www.npmjs.com/package/@motionpicture/sskts-domain)をアップデート。
+
 ## v4.0.2- 2017-11-01
 ### Changed
 - COA本予約時に渡す電話番号から数字以外の文字列を除去。